feat. 모여라 생성 개발중

+++ src/main/java/com/mwom/moyeora/database/dto/MoyeoraDto.java
@@ -1,5 +1,7 @@
 package com.mwom.moyeora.database.dto;
 
+import com.mwom.moyeora.database.entity.MoyeoraEntity;
+import lombok.Builder;
 import lombok.Data;
 
 @Data
@@ -10,6 +12,6 @@ public class MoyeoraDto {
     private String myrTitle;
     private String myrTags;
     private String myrMainImg;
-    private long myrMaxMember;
+    private int myrMaxMember;
 
 }
